+++ docs/src/changelog.md
@@ -13,5 +13,7 @@ Release notes:
 
 ### Added
 
-- none yet
+- `plot_TS_isopycnals` and `plot_TS_spiciness`, to add density and spiciness
+contours to an existing TS plot. These are now used by `plot_TS`, but it can be
+convenient for users to add contours for themselves.